+++ cli/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+
+# 0.5.1
 ### Features
 * Add instructions for users on what to do if no Kurtosis engine is running
 * If an engine isn't running, the CLI will try to start one automatically